refactor: put file ref mgr to sep file
