+++ src/mito2/src/compaction/compactor.rs
@@ -46,7 +46,8 @@ use crate::region::version::VersionRef;
 use crate::region::{ManifestContext, RegionLeaderState, RegionRoleState};
 use crate::schedule::scheduler::LocalScheduler;
 use crate::sst::file::FileMeta;
-use crate::sst::file_purger::{FileReferenceManager, LocalFilePurger};
+use crate::sst::file_purger::LocalFilePurger;
+use crate::sst::file_ref::FileReferenceManager;
 use crate::sst::index::intermediate::IntermediateManager;
 use crate::sst::index::puffin_manager::PuffinManagerFactory;
 use crate::sst::location::region_dir_from_table_dir;
