+++ src/mito2/src/engine.rs
@@ -113,7 +113,7 @@ use crate::read::stream::ScanBatchStream;
 use crate::region::MitoRegionRef;
 use crate::request::{RegionEditRequest, WorkerRequest};
 use crate::sst::file::FileMeta;
-use crate::sst::file_purger::FileReferenceManagerRef;
+use crate::sst::file_ref::FileReferenceManagerRef;
 use crate::wal::entry_distributor::{
     build_wal_entry_distributor_and_receivers, DEFAULT_ENTRY_RECEIVER_BUFFER_SIZE,
 };
