+++ src/mito2/src/region/opener.rs
@@ -60,7 +60,8 @@ use crate::region::{
 use crate::region_write_ctx::RegionWriteCtx;
 use crate::request::OptionOutputTx;
 use crate::schedule::scheduler::SchedulerRef;
-use crate::sst::file_purger::{FileReferenceManagerRef, LocalFilePurger};
+use crate::sst::file_purger::LocalFilePurger;
+use crate::sst::file_ref::FileReferenceManagerRef;
 use crate::sst::index::intermediate::IntermediateManager;
 use crate::sst::index::puffin_manager::PuffinManagerFactory;
 use crate::sst::location::region_dir_from_table_dir;
